Separates Red-Black tree tests over separate topics/modules.
